# Optimize CPU PoS