--- a/crates/shared/ab-proof-of-space/src/chia.rs
+++ b/crates/shared/ab-proof-of-space/src/chia.rs
@@ -26,7 +26,7 @@ impl TableGenerator<ChiaTable> for ChiaTableGenerator {
         }
     }
 
-    #[cfg(any(feature = "parallel", test))]
+    #[cfg(feature = "parallel")]
     fn generate_parallel(&mut self, seed: &PosSeed) -> ChiaTable {
         ChiaTable {
             tables: Tables::<K>::create_parallel((*seed).into(), &self.tables_cache),
@@ -63,7 +63,7 @@ impl Table for ChiaTable {
         }
     }
 
-    #[cfg(all(feature = "alloc", any(feature = "parallel", test)))]
+    #[cfg(feature = "parallel")]
     fn generate_parallel(seed: &PosSeed) -> ChiaTable {
         Self {
             tables: Tables::<K>::create_parallel((*seed).into(), &TablesCache::default()),
@@ -102,14 +102,17 @@ mod tests {
         ]);
 
         let table = ChiaTable::generate(&seed);
+        #[cfg(feature = "parallel")]
         let table_parallel = ChiaTable::generate_parallel(&seed);
 
         assert!(table.find_proof(1232460437).is_none());
+        #[cfg(feature = "parallel")]
         assert!(table_parallel.find_proof(1232460437).is_none());
 
         {
             let challenge_index = 600426542;
             let proof = table.find_proof(challenge_index).unwrap();
+            #[cfg(feature = "parallel")]
             assert_eq!(proof, table_parallel.find_proof(challenge_index).unwrap());
             assert!(ChiaTable::is_proof_valid(&seed, challenge_index, &proof));
         }