--- a/crates/shared/ab-proof-of-space/src/chiapos.rs
+++ b/crates/shared/ab-proof-of-space/src/chiapos.rs
@@ -30,30 +30,20 @@ macro_rules! impl_any {
         $(
 impl Tables<$k> {
     /// Create Chia proof of space tables. There also exists [`Self::create_parallel()`] that trades
-    /// CPU efficiency and memory usage for lower latency.
-    ///
-    /// Advanced version of [`Self::create_simple`] that allows to reuse cache.
-    pub fn create(seed: Seed, cache: &mut TablesCache<$k>) -> Self {
+    /// memory usage for lower latency and higher CPU efficiency.
+    pub fn create(seed: Seed, cache: &TablesCache) -> Self {
         Self(TablesGeneric::<$k>::create(
             seed, cache,
         ))
     }
 
     /// Almost the same as [`Self::create()`], but uses parallelism internally for better
-    /// performance (though not efficiency of CPU and memory usage), if you create multiple tables
-    /// in parallel, prefer [`Self::create()`] for better overall performance.
+    /// latency and performance (though higher memory usage).
     #[cfg(any(feature = "parallel", test))]
-    pub fn create_parallel(seed: Seed, cache: &mut TablesCache<$k>) -> Self {
+    pub fn create_parallel(seed: Seed, cache: &TablesCache) -> Self {
         Self(TablesGeneric::<$k>::create_parallel(
             seed, cache,
         ))
-    }
-
-    /// Create Chia proof of space tables.
-    ///
-    /// Simpler version of [`Self::create`].
-    pub fn create_simple(seed: Seed) -> Self {
-        Self::create(seed, &mut TablesCache::default())
     }
 
     /// Find proof of space quality for given challenge.