A bunch of optimizations here:
* better cleanup for unused memory using new `PrunedTable` abstraction and some fixes
* faster proof verification without heap allocations
* refactored features in `ab-proof-of-space` and fixes to make crate actually usable without `alloc`
* `LeftTargets` is now reused across multiple table generator instances
* table generator instances take `&self` rather than `&mut self` and are cheap to clone, which makes them much easier to use and doesn't require locks for shared access

Verification performance change:
```
Before:
chia/verification       time:   [7.0819 µs 7.0908 µs 7.0961 µs]
                        thrpt:  [140.92 Kelem/s 141.03 Kelem/s 141.21 Kelem/s]
After:
chia/verification       time:   [6.4624 µs 6.4689 µs 6.4758 µs]
                        thrpt:  [154.42 Kelem/s 154.59 Kelem/s 154.74 Kelem/s]
```